Passing in all tests including in mobile devices.
The window.c file used in tests was updated to detect correctly mouse
positions when forcing landscape mode. The test for animated
interfaces is not failing anymore. We corrected a comparison bug and
also included the restriction that the absolute time passed to the
rendering function must be positive. This is done because we interpret
a stored zero measure as meaning that the previous timings are not
initialized. This was making animations delay 1 frame its first
update.